changed hash value to unsigned long

+++ SymbolTable/ScopeTable.h
@@ -13,7 +13,7 @@ class ScopeTable{
         string id;
         // int (*hashFunction)(string name);
         int hashFunction(string name);
-        unsigned long long sdbm(string str);
+        unsigned long sdbm(string str);
     public:
         ScopeTable(int size,string id,ScopeTable * parent);
         bool insert(string name,string type);
